Conflicts:
	firmware/os/driver/ata.c
	firmware/os/driver/ds17485.c
	firmware/os/driver/encx24j600.h
	firmware/os/global.mk

+++ firmware/os/kernel/ksym.c
@@ -7,7 +7,7 @@
 	(c) Stuart Wallace <stuartw@atom.net>, September 2015.
 */
 
-#include <include/limits.h>
+#include <kernel/include/limits.h>
 #include <kernel/ksym.h>
 #include <klibc/stdio.h>
 #include <klibc/string.h>
@@ -72,7 +72,7 @@ s32 ksym_format_nearest_prev(void *addr, char *buf, u32 buf_len)
         if(ent->addr == addr)
             snprintf(buf, buf_len, "<%s>", &ent->name);
         else
-            snprintf(buf, buf_len, "<%s+0x%x>", &ent->name, addr - ent->addr);
+            snprintf(buf, buf_len, "<%s+0x%x>", &ent->name, (u32) addr - (u32) ent->addr);
 
         return SUCCESS;
     }
