+++ firmware/os/platform/lambda_rev0/device.c
@@ -4,17 +4,17 @@
     Stuart Wallace, September 2015
 */
 
-#include <platform/platform.h>
+#include <kernel/platform.h>
 
 #include <stdio.h>
 #include <string.h>
-#include <kutil/kutil.h>
-#include <memory/kmalloc.h>
+#include <kernel/device/auto.h>
+#include <kernel/memory/kmalloc.h>
+#include <kernel/util/kutil.h>
 #include <platform/lambda_rev0/device.h>
-#include <device/auto.h>
-#include <device/mc68681.h>                 /* DUART            */
-#include <device/ds17485.h>                 /* RTC              */
-#include <device/ata.h>                     /* ATA interface    */
+#include <driver/mc68681.h>                 /* DUART            */
+#include <driver/ds17485.h>                 /* RTC              */
+#include <driver/ata.h>                     /* ATA interface    */
 
 
 dev_t *g_lambda_duart;      /* DUART device - stored separately for early console init  */
@@ -117,8 +117,8 @@ void expansion_init()
             mc68681_set_op_bits(g_lambda_console, BIT(EXP_ID));
 
 
-            printf("slot %d (%x-%x, irq %u): ", i, base_addr,
-                    base_addr + EXP_ADDR_LEN - 1, irql);
+            printf("slot %d (%x-%x, irq %u): ", i, (u32) base_addr,
+                    (u32) base_addr + EXP_ADDR_LEN - 1, irql);
 
             ret = dev_auto_init(id, base_addr, irql, NULL, &dev);
